requirements.txt
langchain
langchain-community
sentence_transformers
torch
pypdf
qdrant-client
python-dotenv